requirements.txt
tornado==2.1.1
sockjs-tornado==0.0.4
django==1.4.1
